+++ src/hedera_py_lite/__init__.py
@@ -2,7 +2,7 @@
 
 from hedera_py_lite.client import HederaClient
 
-__version__ = "0.1.0"
+__version__ = "0.2.0"
 __author__ = "Emmanuel Okechukwu Nwajari"
 
 __all__ = ["HederaClient", "__version__"]